Success!
Using a simple Bayer matrix ordered dithering approach yields results
that are definitely good enough. If it mattered, I could follow up with
a void-and-cluster or a full-on error-diffusion dither, but that
probably won't be necessary.

The only downside is that it's slightly less tunable than more
sophisticated approaches, but I can work around that by tuning the
function magnitudes and size of the Bayer matrix.